InvoiceDetailView: Custom Invoice Not Found page
I've added a custom descriptive 404 page for invoices, because it could
be confusing for users with multiple accounts why their links to
invoices don't work.

For security reasons, 404 Not Found is returned when user tries to
access invoice that's not associated with his/hers account, even though
it's technically 403 Forbidden. This is to prevent probes about whether
a specific invoice exists.

This patch doesn't change this behavior, it only provides hints for
people who are trying to download their invoice but can't because
they're logged in a wrong account.