Get rid of confusing example/sample_plans/tests.py
in favor of new example/test_runner.py which runs all tests for plans
module by default without the need to import them explicitly.

This is a rough equivalent of changing the default of

    python manage.py test

to

    python manage.py test plans

Which seems to be a correct way to invoke tests from demo/

Previous behavior with example/sample_plans/tests.py was confusing or
even harmful, as it run a seemingly random subset of tests from
plans/tests/*.py when invoked from demo project.

This change eliminates this confusion and unnecessary duplication.